service/requirements.txt
django==4.0
django-money==2.1
djangorestframework==3.13.0
pyyaml==6.0
uritemplate==4.1.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
requests==2.26.0
prospector==1.5.1
isort==5.9.2
astroid==2.9.0
black==21.11b1
bandit==1.7.1
mypy==0.910
pyroma==3.2
pylint==2.12.1
pylint-django==2.4.4
